+++ config/karma.conf.js
@@ -1,3 +1,5 @@
+
+
 var customLaunchers = {
     sl_chrome: {
         base: 'SauceLabs',
@@ -26,7 +28,7 @@ var customLaunchers = {
 
 
 module.exports = function (config) {
-    config.set({
+    var cfg = {
         basePath: '../',
         frameworks: ['jasmine'],
         files: [
@@ -54,7 +56,7 @@ module.exports = function (config) {
             noInfo: true //please don't spam the console when running in karma!
         },
         // possible values: 'dots', 'progress'
-        reporters: ['mocha', 'coverage', 'saucelabs'],
+        reporters: ['mocha', 'coverage'],
         port: 9876,
         colors: true,
 
@@ -69,20 +71,30 @@ module.exports = function (config) {
         // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
         // browsers: ['Firefox'],
 
-        captureTimeout: 60000,
+        captureTimeout: 120000,
         browserDisconnectTimeout : 10000,
         browserDisconnectTolerance : 1,
-        browserNoActivityTimeout : 60000,
+        browserNoActivityTimeout : 120000,
         // Continuous Integration mode
         // if true, Karma captures browsers, runs the tests and exits
         singleRun: true,
-        sauceLabs: {
+
+    };
+    if(process.env.SAUCE_ACCESS_KEY) {
+        // Use saucelabs for CI tests
+        cfg.reporters.push("saucelabs");
+        cfg.sauceLabs = {
             testName: "Ng2Datastore unit tests",
             startConnect: true,
             build: process.env.CI_BUILD_ID || "manual"
-        },
-        customLaunchers: customLaunchers,
-        browsers: Object.keys(customLaunchers)
+        };
+        cfg.customLaunchers = customLaunchers;
+        cfg.browsers = Object.keys(customLaunchers);
+    } else {
+        // Use PhantomJS for local tests
+        cfg.browsers = ["PhantomJS"]
+    }
 
-    });
+    config.set(cfg);
+
 };