+++ src/backends/interface.ts
@@ -6,7 +6,13 @@ export interface IDSBackend {
     create(identifier: any, values: any, params: any): Observable<any>;
     update(identifier: any, values: any, params: any): Observable<any>;
     destroy(identifier: any, params: any): Observable<any>;
+    action(identifier: any, action: string, params: any): Observable<any>;
 }
+
+export interface IDSBackendClass {
+    new(params: any): IDSBackend;
+}
+
 export interface IDSBackendProvider {
     provide(params: any): IDSBackend;
 }