+++ src/backends/rest.ts
@@ -1,11 +1,11 @@
-import {Headers, URLSearchParams, RequestOptions, Http} from "@angular/http";
+import {Headers, URLSearchParams, RequestOptions, Http, Request, RequestMethod} from "@angular/http";
 import {Observable} from "rxjs";
 import {IDSBackend, IDSBackendProvider} from "./interface";
 import {DSJsonParser} from "../parsers/json";
 import {Injectable, OpaqueToken, Inject, Optional} from "@angular/core";
 import {DSJsonRenderer} from "../renderers/json";
 import "rxjs/add/operator/map";
-
+import {capitalize, isString} from "lodash";
 
 export interface IDSRestIdentifier {
     path: string;
@@ -38,6 +38,7 @@ export class DSRestBackend implements IDSBackend {
 
     public retrieve(identifier: IDSRestIdentifier, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
+        options = this._renderer.prepare(options);
         return this._http
             .get(this.getRequestUrl(identifier), options)
             .map((response) => {
@@ -47,6 +48,7 @@ export class DSRestBackend implements IDSBackend {
 
     public list(identifier: IDSRestIdentifier, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
+        options = this._renderer.prepare(options);
         return this._http
             .get(this.getRequestUrl(identifier), options)
             .map((response) => {
@@ -56,6 +58,7 @@ export class DSRestBackend implements IDSBackend {
 
     public create(identifier: IDSRestIdentifier, values: any, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
+        options = this._renderer.prepare(options);
         return this._http
             .post(this.getRequestUrl(identifier), this._renderer.render(values), options)
             .map((response) => {
@@ -65,6 +68,7 @@ export class DSRestBackend implements IDSBackend {
 
     public update(identifier: IDSRestIdentifier, values: any, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
+        options = this._renderer.prepare(options);
         return this._http
             .put(this.getRequestUrl(identifier), this._renderer.render(values), options)
             .map((response) => {
@@ -74,9 +78,29 @@ export class DSRestBackend implements IDSBackend {
 
     public destroy(identifier: IDSRestIdentifier, params: any = {}): Observable<any> {
         let options = this.getRequestOptions(identifier);
+        options = this._renderer.prepare(options);
         return this._http
             .delete(this.getRequestUrl(identifier), options)
             .map((response) => {
+                return this._parser.parse(response);
+            });
+    }
+
+    public action(identifier: IDSRestIdentifier, action: string, params: any = {}): Observable<any> {
+        // Params : string => url
+        // Params : object => {url, body}
+        let options = this.getRequestOptions(identifier);
+        options = this._renderer.prepare(options);
+        options.method = RequestMethod[capitalize(action)];
+        if (isString(params)) {
+            options.url = this.getRequestUrl(identifier) + <string>params;
+        } else {
+            options.url = this.getRequestUrl(identifier) + params.url;
+            options.body = this._renderer.render(params.body);
+        }
+        return this._http.request(options.url, options)
+            .map((response) => {
+                console.log("ACTION : ", response);
                 return this._parser.parse(response);
             });
     }