+++ src/service/interface.ts
@@ -1,5 +1,0 @@
-import {IDSCollection} from "../collection/interface";
-import {IDSModel} from "../model/interface";
-export interface IDSDataService<T extends IDSModel> {
-    collection(params: any): IDSCollection<T>;
-}