+++ CHANGELOG.md
@@ -1,10 +1,25 @@
-<a name="Unreleased"></a>
-# [Unreleased](https://gitlab.com/solidev/ng2datastore/compare/v0.0.0-alpha.2...) (2016-09-04)
+<a name="0.0.0-alpha.4"></a>
+# [0.0.0-alpha.4](https://gitlab.com/jmbarbier/ng2datastore/compare/v0.0.0-alpha.3...v0.0.0-alpha.4) (2016-09-23)
+
+
+### Bug Fixes
+
+* **model:** Add check for existing collection for model.coll.datastore ([13b13c3](https://gitlab.com/jmbarbier/ng2datastore/commit/13b13c3))
+
+
+### Features
+
+* **collection:** Add xxxx_class for collection setup variants ([70645cc](https://gitlab.com/jmbarbier/ng2datastore/commit/70645cc))
+* **collection:** Added datasources property and setup methods for DSRegister ([5155b34](https://gitlab.com/jmbarbier/ng2datastore/commit/5155b34))
+* **collection:** added list to collection ([85d2b6a](https://gitlab.com/jmbarbier/ng2datastore/commit/85d2b6a))
+* **list:** Make lists working ([af1a6d1](https://gitlab.com/jmbarbier/ng2datastore/commit/af1a6d1))
+* **queryset:** Added queryset instances to collection. ([68efdf5](https://gitlab.com/jmbarbier/ng2datastore/commit/68efdf5))
+* **service:** Removed service layer ([6d1a310](https://gitlab.com/jmbarbier/ng2datastore/commit/6d1a310))
 
 
 
 <a name="0.0.0-alpha.2"></a>
-# [0.0.0-alpha.2](https://gitlab.com/solidev/ng2datastore/compare/...v0.0.0-alpha.2) (2016-09-03)
+# [0.0.0-alpha.2](https://gitlab.com/jmbarbier/ng2datastore/compare/...v0.0.0-alpha.2) (2016-09-03)